requirements.txt
discord.py==1.7.2
motor==2.2.0
